# api/openapi.yaml
openapi: 3.0.3
info:
  title: Stalker API
  description: |
    The API specification developed by GruppOne for the Stalker project.
  contact:
    name: GruppOne
    email: gruppone.swe@gmail.com
    url: 'https://github.com/GruppOne'
  license:
    name: GNU GPLv3
    url: 'https://www.gnu.org/licenses/gpl-3.0-standalone.html'
  version: 0.8.1
servers:
  - url: 'http://localhost:11111'
    description: Development server.
  - url: 'https://{environment}.gruppone.tech/v{majorVersion}'
    variables:
      environment:
        enum:
          - api.staging
          - api
        default: api
      majorVersion:
        default: '0'
    description: Staging and Production servers.
...
paths:
...
  '/organization/{organizationId}':
    delete:
      description: Delete the specified organization. Operation reserved only to the administrator.
      operationId: deleteOrganizationById
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultResponse'
      summary: Delete an organization.
      tags:
        - web app
    get:
      description: Get all information about a specific organization INCLUDING ldapConfiguration.
      operationId: getOrganizationById
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: ./model/organization/Organization.v1.yaml
          description: Ok. The response contains all the public information on the specific organization.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultResponse'
      summary: "Read an organization's information."
      tags:
        - both clients
    parameters:
      - $ref: '#/components/parameters/organizationId'
    put:
      description: Send the changed data about a specific organization. Operation reserved only to the administrator.
      operationId: putOrganizationDataById
      requestBody:
        content:
          application/json:
            schema:
              $ref: ./model/organization/Organization.v1.yaml
        required: true
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultResponse'
      summary: "Update an organization's data."
      tags:
        - web app
...
  '/user/{userId}/organization/{organizationId}/connection':
    parameters:
      - $ref: '#/components/parameters/userId'
      - $ref: '#/components/parameters/organizationId'
    delete:
      summary: Delete a user connection to an organization.
      operationId: delete-user-userId-organizations-connection
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultResponse'
      description: Request the disconnection of a user from an organization.
      tags:
        - mobile app
    post:
      description: The request body should have an LDAP property only if the organization is private.
      operationId: post-user-userId-organization-organizationId-connect
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ldapAuthentication:
                  type: object
                  properties:
                    rdn:
                      type: string
                      description: "The user's RDN in his organization's LDAP server."
                    ldapPassword:
                      type: string
                      format: password
                      description: "The user password for this organization's LDAP server."
                  required:
                    - rdn
                    - ldapPassword
      responses:
        '201':
          description: Created
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          content:
            application/json:
              schema:
                properties:
                  message:
                    enum:
                      - The given LDAP credentials are not valid.
                      - The server could not authenticate the user.
                    type: string
                type: object
          description: Unauthorized
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/DefaultResponse'
      summary: Create a request to connect a user to an organization.
      tags:
        - mobile app
  '/user/{userId}/organizations/connections':
    parameters:
      - $ref: '#/components/parameters/userId'
    get:
      description: Get the list of organizations that a user has connected to.
      operationId: get-user-userId-organizations-connections
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  connectedOrganizationsIds:
                    type: array
                    items:
                      $ref: ./model/GenericId.v1.yaml
                required:
                  - connectedOrganizationIds
          description: Ok. The response contains the list of the ids of the organizations the user is connected to.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultResponse'
      summary: Read the list of all organizations a user is connected to.
      tags:
        - mobile app
...
  '/organization/{organizationId}/users/connections':
    get:
      description: Get the list of all users connected to this organization.
      operationId: get-organization-organizationId-users-connected
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  connectedUsers:
                    type: array
                    items:
                      $ref: ./model/user/User.v1.yaml
                required:
                  - connectedUsers
          description: OK
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultResponse'
      summary: Read the list of users connected to an organization.
      tags:
        - web app
    parameters:
      - $ref: '#/components/parameters/organizationId'
  '/organization/{organizationId}/users/inside':
    get:
      operationId: get-organization-organizationId-users-inside
      summary: Read the number of users inside an organization.
      description: Read the number of users currently inside an organization and its places.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: ./model/organization/UsersInsideOrganization.v1.yaml
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultResponse'
      tags:
        - web app
    parameters:
      - $ref: '#/components/parameters/organizationId'
  '/organization/{organizationId}/users/roles':
    parameters:
      - $ref: '#/components/parameters/organizationId'
    get:
      summary: Read all users with roles in this organization.
      tags:
        - web app
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  usersWithRoles:
                    type: array
                    uniqueItems: true
                    items:
                      $ref: ./model/organization/OrganizationRole.v1.yaml
                required:
                  - usersWithRoles
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/DefaultResponse'
      operationId: get-organization-organizationId-users-roles
      description: Get a list of all users with roles in this organization.
  /organizations:
    get:
      description: 'Get the list of all the organizations registered to Stalker, complete with all details EXCEPT private information (ldapConfiguration).'
      operationId: get-organizations
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizations:
                    type: array
                    items:
                      $ref: ./model/organization/Organization.v1.yaml
                required:
                  - organizations
          description: Ok. The response contains the list of organizations registered to Stalker.
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/DefaultResponse'
      summary: Read all organizations.
      tags:
        - both clients
    post:
      description: Request the creation of a new organization.
      operationId: postOrganizations
      requestBody:
        content:
          application/json:
            schema:
              $ref: ./model/organization/OrganizationData.v1.yaml
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    $ref: ./model/GenericId.v1.yaml
                required:
                  - id
          description: Created. The response contains the id of the newly created organization.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/DefaultResponse'
      summary: Create an organization.
      tags:
        - web app
  /user/login:
    post:
      description: This endpoint should not have a security scheme.
      operationId: postUserLoginCredentials
      requestBody:
        content:
          application/json:
            schema:
              $ref: ./model/user/LoginData.v1.yaml
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                type: object
                properties:
                  jwt:
                    $ref: ./model/user/EncodedJwt.v1.yaml
                required:
                  - jwt
          description: Created. The response contains the encoded JWT authorizing the user.
        '401':
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    enum:
                      - The given email/password combination is not valid.
                    type: string
                required:
                  - message
          description: Unauthorized
        default:
          $ref: '#/components/responses/DefaultResponse'
      security: []
      summary: Create a login request for a user.
      tags:
        - both clients
...
components:
  parameters:
    userId:
      in: path
      name: userId
      description: The id of a user.
      required: true
      schema:
        $ref: ./model/GenericId.v1.yaml
    organizationId:
      in: path
      name: organizationId
      description: The id of an organization.
      required: true
      schema:
        $ref: ./model/GenericId.v1.yaml
...
  responses:
    BadRequest:
      description: The server could not process the request due to a client mistake.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                enum:
                  - The server could not process the request due to a client mistake.
            required:
              - message
    Unauthorized:
      description: The server could not authenticate the user.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                enum:
                  - The server could not authenticate the user.
            required:
              - message
    Forbidden:
      description: The server could not authorize the user.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                enum:
                  - The server could not authorize the user.
            required:
              - message
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                enum:
                  - The requested resource does not exist.
                description: Resource not found.
            required:
              - message
    DefaultResponse:
      description: The server encountered an unexpected error.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                enum:
                  - The server encountered an unexpected error.
                description: Unexpected error.
            required:
              - message
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The token is sent through the Authorization header with Bearer schema.
security:
  - JWT: []