# src/main/java/tech/gruppone/stalker/server/controllers/OrganizationsController.java
package tech.gruppone.stalker.server.controllers;

import java.util.List;
import lombok.AccessLevel;
import lombok.AllArgsConstructor;
import lombok.Value;
import lombok.experimental.FieldDefaults;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;
import reactor.core.publisher.Mono;
import tech.gruppone.stalker.server.model.api.OrganizationDataDto;
import tech.gruppone.stalker.server.model.api.OrganizationDto;
import tech.gruppone.stalker.server.services.OrganizationService;

@AllArgsConstructor
@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
@RestController
@RequestMapping("/organizations")
public class OrganizationsController {

  OrganizationService organizationService;

  @GetMapping
  @ResponseStatus(HttpStatus.OK)
  // FIXME refactor this (no more places only place ids)
  public Mono<GetOrganizationsResponse> getOrganizations() {
    return organizationService.findAll().collectList().map(GetOrganizationsResponse::new);
  }

  @Value
  private static class GetOrganizationsResponse {
    List<OrganizationDto> organizations;
  }

  @PostMapping
  @ResponseStatus(HttpStatus.CREATED)
  public Mono<PostOrganizationsResponse> postOrganizations(
      @RequestBody final OrganizationDataDto organizationDataDto) {
    return organizationService.save(organizationDataDto).map(PostOrganizationsResponse::new);
  }

  @Value
  private static class PostOrganizationsResponse {
    long id;
  }
}